Fixed bug on rate-of-change constraints at updates
The time-indexing for imposing the rate of change constraint on the update timestep had a bug in it. It is a bit confusing because there are corner cases such as 1) Adaptation, where we normally want zero rate of change but want some finite upper bound on the change across the update timestep, and 2) Experiments such as SRM termination or looking at a subset of the portfolio where we may want to constrain a control to be uniquely zero, and this constraint may change across the updating timestep.
Both of these cases should now be handled appropriately. Thanks to Dan Visioni for suggesting the (still in testing) SRM termination experiments that caused me to discover this small bug. This bug didn't affect any of the other notebooks, however.

Also corrected the upper bound SRM rate of change to match what is cited in the paper and improved the replication analyses.