Renamed to ClimateMARGO.jl
This was kind of a pain, because I had to rename the package in all of my examples and docs. I should submit a pull request and try to register the name so that I can finalize it and not have to deal with renaming a bunch of things every again.