Starting re-structuring optimize_controls!
Ideally, we would re-use the functions defined in the `Diagnostics` submodule. I need to test JuMP a bit more to figure out the best way to implement this. Can any of the functions be used right out of the box or do they need to be registered? Do they only need to be registered if they are a function of the control variable? Can they be a function of multiple parameters, including a combination of control variables and other (constant) parameters? Can functions w/ operations
like `cumsum` be registered and operate within JuMP?