Robust io paths and bugs in example notebooks
Added a filter on configuration paths to get rid of anything that's not a .json file (e.g. hidden files in the same directory), which was giving me errors because of an old .ipynb_checkpoints subfolder.

Also, fixed a bug in the Drake_2020_MARGO_paper examples where fonsp had left out some backslashes in front of quotation marks which confused jupyter when reading the JSON format.

Note: the tutorial and 1_default_figures.ipynb seem to be working correctly but it looks like some of the other notebooks have been broken by some bug in the last few commits. Will look into it later.