Improved adaptation model and costs
**Adaptation model**
Following Reviewer #1's suggestion, we improve the adaptation model somewhat following DeBruin et al's AD-DICE implementation. In particular, we assume that adaptation reduces damages proportionally, such that the damages used to compute benefits are the *residual damages*. The marginal costs of these adaptive measures increase super-linearly proportionally with the GWP. Based on suggestions from a different anonymous reviewer (from a previous submission at a different journal), we also make the cost proportional to the controlled gross damages. We tune the proportionality constant to match DeBruin's estimate for the damages that correspond to a warming of 3ºC (this reference temperature comes in as an additional input parameter in the economic submodule, Tb=3.0).

These changes are implemented in both the Diagnostic and Optimization modules.

We remove the constraint that forces adaptation to be constant in time, also following DeBruin's implementation.

**Misc**
Also fixed some bugs with the expenditure and budget objective functions (although these are not used in the paper at all).

Also improved the interface for the standard plots of the model state.