Revisions for submission to *Environmental Research Letters* journal (#73)
* Increase cost exponent to 3

Increased to cost exponent from 2 to 3, which brings MARGO's cost
functions more in line with other models.

Also renamed the units file.

* Improved adaptation model and costs

**Adaptation model**
Following Reviewer #1's suggestion, we improve the adaptation model somewhat following DeBruin et al's AD-DICE implementation. In particular, we assume that adaptation reduces damages proportionally, such that the damages used to compute benefits are the *residual damages*. The marginal costs of these adaptive measures increase super-linearly proportionally with the GWP. Based on suggestions from a different anonymous reviewer (from a previous submission at a different journal), we also make the cost proportional to the controlled gross damages. We tune the proportionality constant to match DeBruin's estimate for the damages that correspond to a warming of 3ºC (this reference temperature comes in as an additional input parameter in the economic submodule, Tb=3.0).

These changes are implemented in both the Diagnostic and Optimization modules.

We remove the constraint that forces adaptation to be constant in time, also following DeBruin's implementation.

**Misc**
Also fixed some bugs with the expenditure and budget objective functions (although these are not used in the paper at all).

Also improved the interface for the standard plots of the model state.

* Increment version number to reflect major (breaking) changes

* Updates tests with new results

* Updated default parameter values based on referrees

- Updated total remove costs in line with new cost function.

- Increased mitigation costs to match other estimates in 2100, rather than in 2020

- Increased climate damages according to Frankhauser review

- Increased geoengineering costs following costs of climate damages

* Fixed bugs in economic submodel and optimization

- Forgot to add Tb as paramter in some sub-module constructor methods

- Forgot to index background damages Db[i] in some places of optimization

* Increased default adaptation limit

* Bugfixes

* Back to de Bruin adaptation module

* Slightly thinner lines

* Fixed one-two dim example

* Removed legacy Tb changes

* Fixed default configuration

* Minor update

* Updated hard-coded correct results for tests